fix jenni hanging when not using SASL or password
If jenni isn't configured to use SASL, a CAP LS was still sent, but the
module is disabled so no CAP END is send after recieving CAP * LS from
the server, leading to a hang. This makes it so that if there's no SASL,
jenni doesn't even begin the CAP handshake.