Updated YouTube module to Version 3 of the API
This commit updates the YouTube module to Version 3 of the Google API, and requires the owner of the bot to provide a Google Developer Server API Key, as required by Google.

The functionality was kept as close as possible to the original module, while also including the .yt command for youtube_search (which is in active use in my community), and providing consistency with the rest of the modules by using jenni instead of bot.