bot: pass setattr calls through to wrapper class transparently
Some modules are setting/updating attributes expecting them to be
globally visible but because the modules only see a wrapped version
of the 'jenni' object, all of these updates are lost immediately after
the call to the module's method.

This change makes jenni.foo = bar apply the update to the wrapped
class and thus apply globally.

See: usage of is_authenticated and is_connected in startup.py and sasl.py